Refactor: Redundant module import
